+++ cli/output.py
@@ -25,3 +25,11 @@ def showProblems(problems, showAnswer):
 
     print('#')
     print('#'*80)
+
+def goodbyeSplash():
+    print('')
+    print('#'*80)
+    print('#')
+    print('#   Thank you for using Math Haxor!')
+    print('#')
+    print('#'*80)