+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.0] - 2023.03.20
 ### Added
 - PDF generator
+- Goodbye splash
 
 ### Fixes
 - Refactor HTML generator into a separate module